1Go alloué

+++ cour0/ConsoleApplication1/ConsoleApplication1/MonHeader.hpp
@@ -4,10 +4,10 @@
 
 struct Vec3
 {
-	float x;
-	float y;
-	float z;
-		
+	int x;
+	int y;
+	int z;
+	
 };
 /*typedef struct Vec3
 {
